+++ pytorch_forecasting/layers/__init__.py
@@ -3,6 +3,7 @@ Architectural deep learning layers from `nn.Module`.
 """
 
 from pytorch_forecasting.layers._attention import AttentionLayer, FullAttention
+from pytorch_forecasting.layers._decomposition import SeriesDecomposition
 from pytorch_forecasting.layers._embeddings import (
     DataEmbedding_inverted,
     EnEmbedding,
@@ -12,13 +13,14 @@ from pytorch_forecasting.layers._encoders import (
     Encoder,
     EncoderLayer,
 )
+from pytorch_forecasting.layers._m_lstm import mLSTMCell, mLSTMLayer, mLSTMNetwork
 from pytorch_forecasting.layers._output._flatten_head import (
     FlattenHead,
 )
+from pytorch_forecasting.layers._s_lstm import sLSTMCell, sLSTMLayer, sLSTMNetwork
 
 __all__ = [
     "FullAttention",
-    "TriangularCausalMask",
     "AttentionLayer",
     "DataEmbedding_inverted",
     "EnEmbedding",
@@ -26,4 +28,11 @@ __all__ = [
     "Encoder",
     "EncoderLayer",
     "FlattenHead",
+    "mLSTMCell",
+    "mLSTMLayer",
+    "mLSTMNetwork",
+    "sLSTMCell",
+    "sLSTMLayer",
+    "sLSTMNetwork",
+    "SeriesDecomposition",
 ]
