+++ pytorch_forecasting/layers/_s_lstm/__init__.py
@@ -1,0 +1,7 @@
+"""sLSTM layer"""
+
+from pytorch_forecasting.layers._s_lstm.cell import sLSTMCell
+from pytorch_forecasting.layers._s_lstm.layer import sLSTMLayer
+from pytorch_forecasting.layers._s_lstm.network import sLSTMNetwork
+
+__all__ = ["sLSTMCell", "sLSTMLayer", "sLSTMNetwork"]
